Handle error where an existing parameter that could be a Standard tier is configured as Advanced tier
This avoids the script throwing
```
botocore.errorfactory.ValidationException: An error occurred (ValidationException) when calling the PutParameter operation: This parameter uses the advanced-parameter tier. You can't downgrade a parameter from the advanced-parameter tier to the standard-parameter tier. If necessary, you can delete the advanced parameter and recreate it as a standard parameter. Be aware that standard parameters have a value limit of 4096 characters.
```